+++ lib/datadog/statsd.rb
@@ -1,4 +1,5 @@
 require 'concurrent'
+require 'monitor'
 require 'socket'
 
 # = Datadog::Statsd: A DogStatsd client (https://www.datadoghq.com)
@@ -86,6 +87,26 @@ module Datadog
     # deprecated, but cannot be removed since uses might use it to check the version against older releases
     def self.VERSION
       VERSION
+    end
+
+    ##
+    # Get the current instance of a StatsD client
+    #
+    # @param *args [Array<Object>]
+    #     Any number of arguments passed to the constructor when building a new instance
+    #
+    # @return [Datadog::Statsd]
+    #     An instance of the StatsD client or a new one using the provided arguments
+    def self.current(*args)
+      @current ||= Datadog::Statsd.new(*args)
+    end
+
+    ##
+    # Set the current StatsD client to be used
+    #
+    # @param client [Datadog::Statsd] The client used on subsequent calls to .current
+    def self.current=(client)
+      @current = client
     end
 
     # @param [String] host your statsd host