+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
  * Try to make this thread safe
+ * Introduce Datadog::Statsd.current to simplify access to a universal client instance
 
 ## 3.3.0 / 2018.02.04